Split context timeline regression tests
