+++ crates/tracepilot-core/src/context_window/tests.rs
@@ -9,6 +9,9 @@ use crate::parsing::events::{RawEvent, TypedEvent, TypedEventData};
 use chrono::Utc;
 use serde_json::json;
 
+mod ownership;
+mod system;
+
 fn event(event_type: SessionEventType, typed_data: TypedEventData) -> TypedEvent {
     TypedEvent {
         raw: RawEvent {
@@ -67,127 +70,6 @@ fn user_message(content: &str, interaction_id: &str) -> TypedEvent {
             parent_agent_task_id: None,
         }),
     )
-}
-
-#[test]
-fn estimates_initial_system_prompt_before_observed_telemetry() {
-    let events = vec![
-        event(
-            SessionEventType::SystemMessage,
-            TypedEventData::SystemMessage(SystemMessageData {
-                content: Some("1234567890abcdef".into()),
-                role: Some("system".into()),
-                name: None,
-                metadata: None,
-            }),
-        ),
-        user_message("abcd", "interaction-1"),
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("1".into()),
-                interaction_id: Some("interaction-1".into()),
-            }),
-        ),
-        assistant_message("efgh"),
-        // A later root prompt replaces the System layer for the next request;
-        // it is neither discarded nor added to Conversation.
-        event(
-            SessionEventType::SystemMessage,
-            TypedEventData::SystemMessage(SystemMessageData {
-                content: Some("1234567890abcdefghijklmnopqrstuv".into()),
-                role: Some("system".into()),
-                name: None,
-                metadata: None,
-            }),
-        ),
-        user_message("ijkl", "interaction-2"),
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("2".into()),
-                interaction_id: Some("interaction-2".into()),
-            }),
-        ),
-        assistant_message("mnop"),
-    ];
-
-    let timeline = build_context_timeline(&events);
-    assert_eq!(timeline.points.len(), 2);
-    assert_eq!(timeline.points[0].system_tokens, 4);
-    assert_eq!(timeline.points[1].system_tokens, 8);
-    assert_eq!(timeline.points[0].tool_definition_tokens, 0);
-    assert_eq!(timeline.points[0].conversation_tokens, 2);
-    assert_eq!(timeline.points[1].conversation_tokens, 4);
-    assert_eq!(timeline.points[0].total_tokens, 6);
-    assert_eq!(timeline.points[1].total_tokens, 12);
-    assert_eq!(timeline.points[0].source, ContextPointSource::Estimated);
-}
-
-#[test]
-fn observed_system_anchor_calibrates_following_prompt_snapshots() {
-    let system_message = || {
-        event(
-            SessionEventType::SystemMessage,
-            TypedEventData::SystemMessage(SystemMessageData {
-                content: Some("1234567890abcdef".into()),
-                role: Some("system".into()),
-                name: None,
-                metadata: None,
-            }),
-        )
-    };
-    let events = vec![
-        system_message(),
-        user_message("abcd", "interaction-1"),
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("1".into()),
-                interaction_id: Some("interaction-1".into()),
-            }),
-        ),
-        assistant_message("efgh"),
-        event(
-            SessionEventType::SessionShutdown,
-            TypedEventData::SessionShutdown(ShutdownData {
-                shutdown_type: None,
-                error_reason: None,
-                total_premium_requests: None,
-                total_api_duration_ms: None,
-                session_start_time: None,
-                events_file_size_bytes: None,
-                current_model: None,
-                current_tokens: Some(12),
-                system_tokens: Some(10),
-                conversation_tokens: Some(2),
-                tool_definitions_tokens: Some(0),
-                total_nano_aiu: None,
-                source_metrics_scope: None,
-                token_details: None,
-                code_changes: None,
-                model_metrics: None,
-                session_segments: None,
-            }),
-        ),
-        system_message(),
-        user_message("ijkl", "interaction-2"),
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("2".into()),
-                interaction_id: Some("interaction-2".into()),
-            }),
-        ),
-        assistant_message("mnop"),
-    ];
-
-    let timeline = build_context_timeline(&events);
-    assert_eq!(timeline.points.len(), 2);
-    assert_eq!(timeline.points[0].source, ContextPointSource::Observed);
-    assert_eq!(timeline.points[0].system_tokens, 10);
-    assert_eq!(timeline.points[1].source, ContextPointSource::Estimated);
-    assert_eq!(timeline.points[1].system_tokens, 10);
 }
 
 #[test]
@@ -347,146 +229,6 @@ fn aligns_tool_contributions_with_reconstructed_conversation_turns() {
     );
     assert_eq!(timeline.top_tool_calls[0].turn, 2);
     assert_eq!(timeline.top_tool_calls[0].result_tokens, 1);
-}
-
-#[test]
-fn excludes_nested_subagent_work_from_main_context_contributions() {
-    let large_child_payload = "subagent-only payload ".repeat(100);
-    let mut child_message = assistant_message(&large_child_payload);
-    if let TypedEventData::AssistantMessage(data) = &mut child_message.typed_data {
-        data.parent_tool_call_id = Some("subagent-call".into());
-    }
-    let mut child_reasoning = event(
-        SessionEventType::AssistantReasoning,
-        TypedEventData::AssistantReasoning(AssistantReasoningData {
-            reasoning_id: Some("child-reasoning".into()),
-            content: Some(large_child_payload.clone()),
-        }),
-    );
-    child_reasoning.raw.agent_id = Some("subagent-call".into());
-
-    let events = vec![
-        user_message("delegate this", "interaction-1"),
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("turn-1".into()),
-                interaction_id: Some("interaction-1".into()),
-            }),
-        ),
-        event(
-            SessionEventType::ToolExecutionStart,
-            TypedEventData::ToolExecutionStart(ToolExecStartData {
-                tool_call_id: Some("subagent-call".into()),
-                turn_id: Some("turn-1".into()),
-                tool_name: Some("task".into()),
-                arguments: Some(json!({"prompt": "review"})),
-                parent_tool_call_id: None,
-                mcp_server_name: None,
-                mcp_tool_name: None,
-            }),
-        ),
-        event(
-            SessionEventType::SubagentStarted,
-            TypedEventData::SubagentStarted(SubagentStartedData {
-                tool_call_id: Some("subagent-call".into()),
-                agent_name: Some("review".into()),
-                agent_display_name: Some("Review Agent".into()),
-                agent_description: None,
-            }),
-        ),
-        child_message,
-        child_reasoning,
-        event(
-            SessionEventType::ToolExecutionStart,
-            TypedEventData::ToolExecutionStart(ToolExecStartData {
-                tool_call_id: Some("child-tool".into()),
-                turn_id: Some("turn-1".into()),
-                tool_name: Some("view".into()),
-                arguments: Some(json!({"payload": large_child_payload})),
-                parent_tool_call_id: Some("subagent-call".into()),
-                mcp_server_name: None,
-                mcp_tool_name: None,
-            }),
-        ),
-        event(
-            SessionEventType::ToolExecutionComplete,
-            TypedEventData::ToolExecutionComplete(ToolExecCompleteData {
-                tool_call_id: Some("child-tool".into()),
-                turn_id: Some("turn-1".into()),
-                parent_tool_call_id: Some("subagent-call".into()),
-                model: Some("gpt-5.6-luna".into()),
-                interaction_id: Some("interaction-1".into()),
-                success: Some(true),
-                result: Some(json!({"content": large_child_payload})),
-                error: None,
-                tool_telemetry: None,
-                is_user_requested: None,
-            }),
-        ),
-        event(
-            SessionEventType::ToolExecutionComplete,
-            TypedEventData::ToolExecutionComplete(ToolExecCompleteData {
-                tool_call_id: Some("subagent-call".into()),
-                turn_id: Some("turn-1".into()),
-                parent_tool_call_id: None,
-                model: Some("gpt-5.6-luna".into()),
-                interaction_id: Some("interaction-1".into()),
-                success: Some(true),
-                result: Some(json!({"content": "review complete"})),
-                error: None,
-                tool_telemetry: None,
-                is_user_requested: None,
-            }),
-        ),
-    ];
-
-    let timeline = build_context_timeline(&events);
-    assert_eq!(timeline.top_tool_calls.len(), 1);
-    assert_eq!(timeline.top_tool_calls[0].tool_name, "task");
-    assert!(
-        timeline.points[0].conversation_tokens < 100,
-        "subagent-private context leaked into the main context estimate"
-    );
-}
-
-#[test]
-fn folded_skill_context_is_counted_once_and_not_shown_as_user_input() {
-    let skill_body = "Follow this skill workflow exactly.";
-    let folded_context = format!("<skill-context name=\"demo\">{skill_body}</skill-context>");
-    let mut invocation = event(
-        SessionEventType::SkillInvoked,
-        TypedEventData::SkillInvoked(SkillInvokedData {
-            name: Some("demo".into()),
-            path: None,
-            content: Some(skill_body.into()),
-            allowed_tools: None,
-            plugin_name: None,
-            plugin_version: None,
-            description: None,
-        }),
-    );
-    invocation.raw.id = Some("skill-event".into());
-    let mut folded_message = user_message(&folded_context, "interaction-1");
-    folded_message.raw.parent_id = Some("skill-event".into());
-
-    let events = vec![
-        invocation,
-        folded_message,
-        event(
-            SessionEventType::AssistantTurnStart,
-            TypedEventData::TurnStart(TurnStartData {
-                turn_id: Some("turn-1".into()),
-                interaction_id: Some("interaction-1".into()),
-            }),
-        ),
-        assistant_message("done"),
-    ];
-
-    let timeline = build_context_timeline(&events);
-    let expected_tokens = (folded_context.len() as u64).div_ceil(4) + 1;
-    assert_eq!(timeline.points[0].conversation_tokens, expected_tokens);
-    assert!(timeline.events.is_empty());
 }
 
 #[test]
